+++ CHANGELOG.md
@@ -12,6 +12,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1109 YourLoops main menu sometimes do not disappear as it should
 - YLP-1116 Missing password strength meter into reset password page
 ### Engineering Use
+- YLP-1082 Improve daily view rendering
 - YLP-1114 Review routes and data structure for message-api
 
 ## 2.0.5 - 2021-12-06